[fix] Link system version of libpcsclite-dev for native targets (#9)
When building for native Linux targets (empty target option), linking against the vendored binaries in `_build/linux` causes Zig (I think) to add a runpath pointing to `_build/linux` - this causes the vendored lib to be linked at runtime instead of the system lib, which in turn causes issues if there's a protocol version mismatch between the vendored client and the system's `pcscd`. 

- Reproduced the issue with a plain `zig build e2e` on a Debian 13 with `libpcsclite1@2.3.3-1` (the vendored version is `1.9.9`) - manifests as a `SericeStopped` error when attempting to establish a new connection.
  - Can be worked around by stripping the runpath (e.g. `patchelf --remove-rpath /path/to/binary`).
  - Could potentially use `std.Build.Module.addRPathSpecial()` to prepend a path to the expected location of the system's `libpcsclite` lib, but want to avoid hardcoding those, even if the pkg-config metadata included with the vendored libs could help here.
- There's no issue when an explicit target is specified (e.g. `zig build e2e -Dtarget=aarch64-linux-gnu`) - the runpath is omitted and the system `libpcsclite1` is linked at runtime, as expected.

Adding a `link_system_pcsclite` user option to work around the issue - when set to `true`, we skip adding the path to the vendored libs as a library search path, ensuring both compile-time and runtime linking are done with the system lib.